+++ packages/d3fc-annotation/index.js
@@ -1,6 +1,7 @@
 export { default as annotationSvgBand } from './src/svg/band';
 export { default as annotationCanvasBand } from './src/canvas/band';
 export { default as annotationSvgCrosshair } from './src/svg/crosshair';
+export { default as annotationCanvasCrosshair } from './src/canvas/crosshair';
 export { default as annotationSvgLine } from './src/svg/line';
 export { default as annotationCanvasLine } from './src/canvas/line';
 export { default as annotationSvgGridline } from './src/svg/gridline';