+++ packages/d3fc-chart/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="2.5.7"></a>
+## [2.5.7](https://github.com/d3fc/d3fc/compare/@d3fc/d3fc-chart@2.5.6...@d3fc/d3fc-chart@2.5.7) (2019-05-01)
+
+
+
+
+**Note:** Version bump only for package @d3fc/d3fc-chart
+
 <a name="2.5.6"></a>
 ## [2.5.6](https://github.com/d3fc/d3fc/compare/@d3fc/d3fc-chart@2.5.5...@d3fc/d3fc-chart@2.5.6) (2019-04-04)